Added extra layer of security into database

+++ includes/db.php
@@ -71,6 +71,9 @@ function db_get_array($db, $sql) {
 
 
 function db_get_user_from_username($db, $username, $include_closed) {
+    // Add slashes into username
+    $username = addslashes($username);
+    
 	// Convert username to lowercase
 	$username = strtolower($username);
 	
@@ -83,6 +86,9 @@ function db_get_user_from_username($db, $username, $include_closed) {
 }
 
 function db_get_user_from_email($db, $email, $include_closed) {
+    // Add slashes into email
+    $email = addslashes($email);
+    
 	// Convert email to lowercase
 	$email = strtolower($email);
     
@@ -95,6 +101,12 @@ function db_get_user_from_email($db, $email, $include_closed) {
 }
 
 function db_get_user_from_session_id($db, $session_id) {
+    // Check that session_id is numeric
+    if (!is_numeric($session_id)) {
+        return null;
+    }
+    
+    // Query
     return db_get_single($db, "SELECT * FROM session, user WHERE session.id=$session_id AND session.user_id=user.id");
 }
 
@@ -103,20 +115,31 @@ function db_get_category_list($db) {
 }
 
 function db_get_category_from_id($db, $category_id) {
+    // Check that category_id is numeric
+    if (!is_numeric($category_id)) {
+        return null;
+    }
+    
+    // Query
     return db_get_single($db, "SELECT * FROM category WHERE id=$category_id");
 }
 
 function db_get_product_list($db, $category_id, $count, $start, $user_id) {
+    // Check that count and start are numeric
+    if (!is_numeric($count) || !is_numeric($start)) {
+        return null;
+    }
+    
     // Build Query
         $sql = "SELECT * FROM product";
         
-        // If user id is set, get basket item too
-        if ($user_id) {
+        // If user id is set and is numeric, get basket item too
+        if ($user_id && is_numeric($user_id)) {
             $sql = $sql." LEFT JOIN basketitem ON basketitem.product_id=product.id AND basketitem.user_id=$user_id";
         }
         
         // If category id is set, only lookup values with that category
-        if ($category_id) {
+        if ($category_id && is_numeric($category_id)) {
             $sql = $sql." WHERE category_id=$category_id";
         }
         
@@ -128,12 +151,17 @@ function db_get_product_list($db, $category_id, $count, $start, $user_id) {
 }
 
 function db_get_product_from_id($db, $product_id, $user_id) {
+    // Check that product_id is numeric
+    if (!is_numeric($product_id)) {
+        return null;
+    }
+    
     // If user was provided, look for basket item as well
     $sql = null;
-    if ($user_id) {
+    if ($user_id && is_numeric($user_id)) {
     	$sql = "SELECT * FROM product LEFT JOIN basketitem ON basketitem.product_id=product.id AND basketitem.user_id=$user_id WHERE product.id=$product_id";
     } else {
-    	$sql = "SELECT * FROM product WHERE id = \"$product_id\"";
+    	$sql = "SELECT * FROM product WHERE id=$product_id";
     }
     
     // Return
@@ -145,6 +173,11 @@ function db_get_basket_from_user_id($db, $user_id) {
 }
 
 function db_set_basket_item($db, $user_id, $product_id, $quantity) {
+    // Check that user_id, product_id and quantity are all numeric
+    if (!is_numeric($user_id) || !is_numeric($product_id) || !is_numeric($quantity)) {
+        return;
+    }
+    
     // If quantity is 0, delete the basket item
     if ($quantity == 0) {
         $db->query("DELETE FROM basketitem WHERE user_id=$user_id AND product_id=$product_id");
@@ -155,6 +188,20 @@ function db_set_basket_item($db, $user_id, $product_id, $quantity) {
 }
 
 function db_place_order($db, $user_id, $shipping_cost, $vat, $name, $address1, $address2, $town, $county, $postcode, $phone) {
+    // Check that user_id, shipping_cost and vat are numeric
+    if (!is_numeric($user_id) || !is_numeric($shipping_cost) || !is_numeric($vat)) {
+        return;
+    }
+    
+    // Add slashes into all address fields
+    $name = addslashes($name);
+    $address1 = addslashes($address1);
+    $address2 = addslashes($address2);
+    $town = addslashes($town);
+    $county = addslashes($county);
+    $postcode = addslashes($postcode);
+    $phone = addslashes($phone);
+    
     // Since quite a lot is going on, we will need to use MySQL transactions
     // to make sure if there is a problem, the database doesn't get messed up
